Generate code for q3 and Change content file q3_memory to q3_time
